Update README.md
Balance Accounting provides you with
respectable accountancy and bookkeeping
services that you demand. The services we
offer meet a wide range of both corporate and
personal needs. After working with us, you’ll be
able to leave your accounting worries behind
and get back to focusing on the things that
matter most in your life.